feat: improve mdx component usage, adding props table and heading to main view
